Add files via upload
Shopping cars (maintenance events) is a large logistical problem for rail car owners and managers. One pain point is not knowing how long a car will be out of service while in a repair facility (shop). This project comprises creating a regressor to predict how long a car will be in the shop (Cycle Time).